WIP fragments..

+++ src/resolvers/Mutation.js
@@ -51,25 +51,24 @@ const Mutation = {
       token: jwt.sign({ userId: user.id }, 'thisismysecretkey') //Create a new JWT token.
     }
   },
-  async deleteUser(parent, args, { prisma }, info) {
-    const userExists = await prisma.exists.User({ id: args.id });
-    if(!userExists) {
-      throw new Error('User not found.');
-    }
+  async deleteUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.deleteUser({
         where: {
-          id: args.id
+          id: userId
         }
       }, info);
   },
-  updateUser(parent, args, { prisma }, info) {
-    // You can do some validation like above, or just let prisma do it.
+  updateUser(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.updateUser({
       where: {
-        id: args.id
+        id: userId
       },
       data: args.data
-    });
+    }, info);
   },
   createPost(parent, args, { prisma, request }, info) {
     const userId = getUserId(request);
@@ -87,14 +86,32 @@ const Mutation = {
       }
     }, info);
   },
-  deletePost(parent, args, { prisma }, info) {
+  async deletePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if(!postExists) throw new Error('Unable to delete post');
+
     return prisma.mutation.deletePost({
       where: {
         id: args.id
       }
     }, info);
   },
-  updatePost(parent, args, { prisma }, info) {
+  async updatePost(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const postExists = await prisma.exists.Post({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if(!postExists) throw new Error('Unable to update post');
+
     return prisma.mutation.updatePost({
       data: args.data,
       where: {
@@ -102,7 +119,9 @@ const Mutation = {
       }
     }, info);
   },
-  createComment(parent, args, { prisma }, info) {
+  createComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+
     return prisma.mutation.createComment({
       data: {
         text: args.data.text,
@@ -113,20 +132,36 @@ const Mutation = {
         },
         author: {
           connect: {
-            id: args.data.author
+            id: userId
           }
         }
       }
     }, info);
   },
-  deleteComment(parent, args, { prisma }, info) {
+  async deleteComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if(!commentExists) throw new Error('Unable to delete comment');
     return prisma.mutation.deleteComment({
       where: {
         id: args.id
       }
     }, info);
   },
-  updateComment(parent, args, { prisma }, info) {
+  async updateComment(parent, args, { prisma, request }, info) {
+    const userId = getUserId(request);
+    const commentExists = await prisma.exists.Comment({
+      id: args.id,
+      author: {
+        id: userId
+      }
+    });
+    if(!commentExists) throw new Error('Unable to update comment');
     return prisma.mutation.updateComment({
       where: {
         id: args.id
